+++ src/Renderers/Tubes/TriangleTubesCPU.cpp
@@ -39,7 +39,7 @@ void createTriangleTubesRenderDataCPU(
         uint32_t linePointOffset,
         std::vector<glm::vec3>& lineTangents,
         std::vector<glm::vec3>& lineNormals) {
-    if (numCircleSubdivisions != globalCircleVertexPositions.size() || tubeRadius != globalTubeRadius) {
+    if (size_t(numCircleSubdivisions) != globalCircleVertexPositions.size() || tubeRadius != globalTubeRadius) {
         initGlobalCircleVertexPositions(numCircleSubdivisions, tubeRadius);
     }
 
@@ -127,7 +127,8 @@ void createTriangleEllipticTubesRenderDataCPU(
         uint32_t linePointOffset,
         std::vector<glm::vec3>& lineTangents,
         std::vector<glm::vec3>& lineNormals) {
-    if (numCircleSubdivisions != globalEllipseVertexPositions.size() || tubeNormalRadius != globalTubeNormalRadius
+    if (size_t(numCircleSubdivisions) != globalEllipseVertexPositions.size()
+            || tubeNormalRadius != globalTubeNormalRadius
             || tubeBinormalRadius != globalTubeBinormalRadius) {
         initGlobalEllipseVertexPositions(numCircleSubdivisions, tubeNormalRadius, tubeBinormalRadius);
     }
